+++ shared/src/main/scala/ky/korins/blake3/HasherImpl.scala
@@ -165,6 +165,15 @@ private[blake3] class HasherImpl(
   def done(): Byte =
     getOutput.rootByte()
 
+  def doneShort(): Short =
+    getOutput.rootShort()
+
+  def doneInt(): Int =
+    getOutput.rootInt()
+
+  def doneLong(): Long =
+    getOutput.rootLong()
+
   def done(out: OutputStream, len: Int): Unit =
     getOutput.rootBytes(out, len)